+++ models/events.go
@@ -3,9 +3,10 @@ package models
 import (
 	"context"
 
+	"github.com/nyaruka/goflow/flows"
+
 	"github.com/gomodule/redigo/redis"
 	"github.com/jmoiron/sqlx"
-	"github.com/nyaruka/goflow/flows"
 	"github.com/pkg/errors"
 )